Move platform independent RenderService and Cache classes to the Mapsui project
