+++ Mapsui/Rendering/Caching/CacheHolder.cs
