+++ Mapsui/Rendering/Caching/CacheTracker.cs
