+++ Mapsui/Rendering/Caching/DrawableImageCache.cs
