+++ Mapsui/Rendering/Caching/ICacheHolder.cs
