+++ Mapsui/Rendering/Caching/ITileCacheEntry.cs
