+++ Mapsui/Rendering/Caching/TileCache.cs
