+++ Mapsui/Rendering/Caching/VectorCache.cs
